==> 03_logistic_1.R <==
#WARNING WARNING WARNING WARNING!
#WARNING WARNING WARNING WARNING!
#WARNING WARNING WARNING WARNING!
#WARNING WARNING WARNING WARNING!
#This code takes about 8 hours to fully run, and uses significant RAM.
###########################################################################

#Here we perform logistic regression on the dataset,
#using first a training/holdout split to identify the
#best candidate features, then using stepwise variable selection to
#refine the best candidates, and then 5-fold CV to ultimately evaluate their performance.o

#The R environment is saved after the training/holdout comparisons are made.
#However, the environment after CV evaluation is very, very large
#As such, the CV results generated by the vglm_evaluate_cv function
#are saved as individual rds files. 

#vglm objects are not saved, as there are too many and they are too large.
#However, all performance measures for all models are saved, including:
#misclassification rate
#per class misclassification rate
#50 and 80% prediction interval frequencies and average sizes
#models selected after stepwise model selection.

#We do not print out any results in this file, for code brevity and clarity
#However, all results are available in the individual result variables/rds files
#generated by the evaluation functions

#See 01_functions.R for descriptions of the "performance" function, and the 
#"kfold_cv" function
set.seed(12345)
library(tidyverse)
library(VGAM)

#Load data and accessory function, stablish a training/holdout set
load("functions.RData")
data = read_rds("transformed_dataset.rds")

sample_indices=sample(1:nrow(data),nrow(data)/2)
training=data[sample_indices,]
holdout=data[setdiff(1:nrow(data),sample_indices),]

#used for clarity when using the evaluation functions
responsename="Cover_Type"

#' Function for fitting a multinomial logistic model to a training set, and 
#' then evaluating its performance on a holdout set.Requires the VGAM library
#' @description Function for fitting a multinomial logistic model to a training set,
#' and then evaluating its performance on a holdout set.
#' @return The output of the performance function, detailed in the 01_functions.R file
#' This includes 50 and 80% interval average sizes and frequencies and misclassification errors
#' from predictions on the holdout set.
#' @param formula The regression formula used to fit the multinomial logistic regression model
#' @param training the dataset used to fit the model
#' @param holdout the holdout dataset used to evaluate the out of sample performance of the model fitted
#' on the training set
#' @param responsename The name of the response variable in the dataset
vglm_evaluate <- function(formula,training,holdout,responsename){
  model = vglm(formula,family=multinomial(),data=training)
  #prediction probabilities
  holdout_pmatrix = predict(model, newdata=holdout, type="response")
  #keep the memory usage down (hopefully)
  rm(model)
  #See 01_functions.R for the output of the performance function
  return(performance(holdout_pmatrix,unlist(holdout[,responsename])))
}

#' A function for performing stepwise variable selection on a multinomial logistic model, and then 
#' evaluating evaluating this chosen model using 5-fold cross validation. This includes 50 and 80%
#' prediction interval statistics, and misclassification rates for each class.
#' Also fits this same chosen model to a training set and evaluates performance
#' measures on a holdout set, in order to generate 50 and 80 percent classification prediction
#' interval frequencies. Requires the VGAM library
#' @description Performs stepwise variable selection on a multinomial logistic regression model,
#' and then evaluates its performing using 5-fold cross validation, 
#' @return A list containing five components:
#' cv_results: the performance results of 5-fold cross validation, see the kfold_cv function documentation
#' holdout_results: the performance results using a training/holdout split, see the performance function documentation
#' selected_formula: the formula selected after performing stepwise variable selection with BIC as the selection criteria
#' selected_steps: the steps taken by the stepwise variables selection
#' coefficients: the coefficients of the fitted multinomial logistic model
#' @param formula The formula for the initial multinomial logistic regression model
#' @param data the complete dataset for 5-fold cross validation
#' @param training the training set, a subset of data, used to fit a model for training/holdout performance evaluation
#' @param holdout the holdout set, for training/holdout performance evaluation
#' @param responsename the name of the response variable

vglm_evaluate_cv <- function(formula,data,training,holdout,responsename){
  model = vglm(formula,family=multinomial(),data=data)
  n = nrow(data)
  #Perform stepwise variable selection with BIC
  selected_model=step4vglm(model,k=log(n)) #
  rm(model) #Reduce memory wastage
  #Theformula selected by stepwise variable selection
  selected_formula=as.formula(as.character(selected_model@call)[2])
  #The steps used in stepwise variable selection
  selected_steps = selected_model@post$anova
  #the coefficients of the selected model
  coefficients = selected_model@coefficients
  rm(selected_model)
  
  #apply training/holdout evaluation in order to generate prediction interval frequencies
  #see documentation for performance function in 01_functions.R
  trained_model = vglm(selected_formula,family=multinomial(),data=training)
  holdout_pmatrix = predict(trained_model, newdata=holdout, type="response")
  holdout_results = performance(holdout_pmatrix,unlist(holdout[,responsename]))
  
  #apply 5-fold cross validation, using the kfold_cv function. See 01_functions.R
  cv_results=kfold_cv(data,vglm,model_formula=selected_formula,pred_type = "response",family=multinomial())
  return(list(
    cv_results = cv_results,
    holdout_results = holdout_results,
    selected_formula=selected_formula,
    selected_steps=selected_steps,
    coefficients = coefficients
  ))
}

#Here we use training holdout splits to identify possible candidates for full CV evaluation,
#and time-costing stepwise variable selection
###########################################################################################################

#Full dataset, with only soil type transformation
full_model = vglm_evaluate(Cover_Type ~ Elevation+
                        Aspect+Slope+Horizontal_Distance_To_Hydrology+
                        Vertical_Distance_To_Hydrology+
                        Horizontal_Distance_To_Roadways+
                        Hillshade_9am+Hillshade_Noon+
                        Hillshade_3pm+Horizontal_Distance_To_Fire_Points+
                        Soil_Type_Dropped+Wilderness_Area,training,holdout,responsename)


#Only elevation, soil type, and wilderness area
elevation_soil_wilderness = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area,
                                          training,holdout,responsename)


#Include log or cube transformation of distance measurements to elevation, soil type, and wilderness area
distances_log = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                           log_Horizontal_Distance_To_Roadways+
                           log_Horizontal_Distance_To_Hydrology+
                           log_Horizontal_Distance_To_Fire_Points+
                           cube_Vertical_Distance_To_Hydrology,training,holdout,responsename)


#Include binned transformation of distance measurements to elevation, soil type, and wilderness area
distances_binned = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                              Horizontal_Distance_To_Roadways_Binned+
                              Horizontal_Distance_To_Hydrology_Binned+
                              Horizontal_Distance_To_Fire_Points_Binned+
                              Vertical_Distance_To_Hydrology_Binned,training,holdout,responsename)


#Include hillshade measurements to elevation, soil type, and wilderness area
hillshades=vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                      Hillshade_9am+
                      Hillshade_Noon+Hillshade_3pm,training,holdout,responsename)


#Include the averaged hillshade measurement (average of 9am, noon, and 3pm measurements) to 
#elevation, soil type, and wilderness area
hillshade_average=vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                             Hillshade_Average,training,holdout,responsename)


#Include binned aspect area measurements to elevation, soil type, and wilderness area
aspect_binned = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Aspect_Binned+
                           Wilderness_Area,training,holdout,responsename)

#Include transformed or "rotated" aspect measurement to elevation, soil type, and wilderness area
aspect_rotated = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Aspect_Transformed+Wilderness_Area,
                                training,holdout,responsename)


#Now we attempt adding suspected interaction terms

#Elevation and soil type interaction
interactions_soil_elevation = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+
                                              Elevation:Soil_Type_Dropped+Wilderness_Area,
                                              training,holdout,responsename)


#Log distance measurements and elevation interaction

interactions_distance_elevation = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                                                  Elevation:log_Horizontal_Distance_To_Hydrology+
                                                  Elevation:log_Horizontal_Distance_To_Roadways + 
                                                  log_Horizontal_Distance_To_Hydrology + 
                                                  log_Horizontal_Distance_To_Roadways,
                                                  training,holdout,responsename)


#Hillshade and binned aspect interaction

interactions_aspect_hillshade = vglm_evaluate( Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                                                 Aspect_Binned+Hillshade_9am+Hillshade_Noon+Hillshade_3pm+
                                                 Aspect_Binned:Hillshade_9am+Aspect_Binned:Hillshade_Noon+
                                                 Aspect_Binned:Hillshade_3pm,
                                                 training,holdout,responsename)



#Hillshade and elevation interaction
interactions_hillshade_elevation = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area +
                                                   Hillshade_9am+Hillshade_Noon+Hillshade_3pm+
                                                   Elevation:Hillshade_9am+Elevation:Hillshade_Noon+
                                                   Elevation:Hillshade_3pm,
                                                   training,holdout,responsename)

#Binned Distance measurement and elevation interaction

interactions_distances_elevation = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                                                   Horizontal_Distance_To_Roadways_Binned+
                                                   Horizontal_Distance_To_Hydrology_Binned+
                                                   Horizontal_Distance_To_Fire_Points_Binned+
                                                   Vertical_Distance_To_Hydrology_Binned+
                                                   Elevation:Horizontal_Distance_To_Roadways_Binned+
                                                   Elevation:Horizontal_Distance_To_Hydrology_Binned+
                                                   Elevation:Horizontal_Distance_To_Fire_Points_Binned+
                                                   Elevation:Vertical_Distance_To_Hydrology_Binned,
                                                   training,holdout,responsename)

#Slope Binned Aspect interaction

slope_interaction = vglm_evaluate(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+Slope+
                                    Elevation:Slope+Aspect_Binned+Aspect_Binned:Slope+
                                    Elevation:Aspect_Binned,
                                    training,holdout,responsename)

save.image("holdout_evaluation.RData")

#Using the previous out of sample prediction error from holdout model evaluations, 
#we apply stepwise variable selection and 5-fold CV to the top candidates models. The full untransformed
#model and a small subset of suspected good predictors (elevation_soil_wilderness) is also included for comparison

#Results of variable selection and CV are saved as RDS files, as the environment is quite large at this point
#######################################################################################################################
#Full dataset, with only soil type transformation
full_model = vglm_evaluate_cv(Cover_Type ~ Elevation+
                             Aspect+Slope+Horizontal_Distance_To_Hydrology+
                             Vertical_Distance_To_Hydrology+
                             Horizontal_Distance_To_Roadways+
                             Hillshade_9am+Hillshade_Noon+
                             Hillshade_3pm+Horizontal_Distance_To_Fire_Points+
                             Soil_Type_Dropped+Wilderness_Area,data,training,holdout,responsename)
write_rds(full_model,"full_model.rds")

#Only elevation, soil type, and wilderness area
elevation_soil_wilderness = vglm_evaluate_cv(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area,data,
                                             training,holdout,responsename)
write_rds(elevation_soil,"elevation_soil_wilderness_area.rds")

#Include binned transformation of distance measurements to elevation, soil type, and wilderness area
distances_binned = vglm_evaluate_cv(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                                   Horizontal_Distance_To_Roadways_Binned+
                                   Horizontal_Distance_To_Hydrology_Binned+
                                   Horizontal_Distance_To_Fire_Points_Binned+
                                   Vertical_Distance_To_Hydrology_Binned,data,training,holdout,responsename)
write_rds(distances_binned,"distances_binned.rds")

#Include hillshade measurements to elevation, soil type, and wilderness area
hillshades=vglm_evaluate_cv(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                           Hillshade_9am+
                           Hillshade_Noon+Hillshade_3pm,data,training,holdout,responsename)
write_rds(hillshades,"hillshades.rds")

#Elevation Binned Distance interaction
interactions_distances_elevation = vglm_evaluate_cv(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                                                   Horizontal_Distance_To_Roadways_Binned+
                                                   Horizontal_Distance_To_Hydrology_Binned+
                                                   Horizontal_Distance_To_Fire_Points_Binned+
                                                   Vertical_Distance_To_Hydrology_Binned+
                                                   Elevation:Horizontal_Distance_To_Roadways_Binned+
                                                   Elevation:Horizontal_Distance_To_Hydrology_Binned+
                                                   Elevation:Horizontal_Distance_To_Fire_Points_Binned+
                                                   Elevation:Vertical_Distance_To_Hydrology_Binned,data,
                                                 training,holdout,responsename)
write_rds(interactions_distances_elevation,"interactions_distances_elevation.rds")

#Hillshade and binned aspect interaction

interactions_aspect_hillshade = vglm_evaluate_cv( Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+
                                                 Aspect_Binned+Hillshade_9am+Hillshade_Noon+Hillshade_3pm+
                                                 Aspect_Binned:Hillshade_9am+Aspect_Binned:Hillshade_Noon+
                                                 Aspect_Binned:Hillshade_3pm,data,
                                               training,holdout,responsename)
write_rds(interactions_aspect_hillshade,"interactions_aspect_hillshade.rds")

slope_interaction = vglm_evaluate_cv(Cover_Type~Elevation+Soil_Type_Dropped+Wilderness_Area+Slope+
                                    Elevation:Slope+Aspect_Binned+Aspect_Binned:Slope+
                                    Elevation:Aspect_Binned,data,
                                  training,holdout,responsename)
write_rds(slope_interaction,"slope_interaction.rds")



#Now we combine the seeming best variables into the best predictor, using all of the best improvements
##########################################################################################################
best_formula = Cover_Type~ Elevation + Soil_Type_Dropped + Wilderness_Area + 
  Slope + Aspect_Binned + Elevation:Slope + Slope:Aspect_Binned +
  Hillshade_9am+Hillshade_Noon+Hillshade_3pm+Horizontal_Distance_To_Roadways_Binned+
  Horizontal_Distance_To_Hydrology_Binned+Horizontal_Distance_To_Fire_Points_Binned+
  Vertical_Distance_To_Hydrology_Binned+Elevation:Horizontal_Distance_To_Roadways_Binned+
  Elevation:Horizontal_Distance_To_Hydrology_Binned+Elevation:Horizontal_Distance_To_Fire_Points_Binned+
  Elevation:Vertical_Distance_To_Hydrology_Binned+Aspect_Binned:Hillshade_9am+Aspect_Binned:Hillshade_Noon+
  Aspect_Binned:Hillshade_3pm

best_model_results=vglm_evaluate_cv(best_formula,data,training,holdout,responsename)
write_rds(best_model_results,"best_model.rds")
